Make plan and matching module siblings
